fix(wit): correct layer-idx from u32 to s32 in world-prepass inline WIT
The prepass world used `type layer-idx = u32` but raft-prefix support
layers carry negative indices (-1, -2, …) and `paint-region-entry.layer-index`
was already typed as `layer-idx` in the hand-rolled record but as plain `u32`
in the inline WIT. The macro glue at slicer-macros/src/lib.rs:1317 adds
`use geometry.{ex-polygon, polygon, point2}` to the inline WIT to satisfy
wit-bindgen 0.24 flat-type-re-export rules — `polygon` and `point2` need
explicit use statements because their alias TypeId lacks TypeInfo.owned/
borrowed, so generate! does not re-export them automatically.

The dispatch.rs cast `entry.layer_index as u32` on harvest was a latent bug:
the validator already rejected negative values, so this cast never triggered,
but the fix is to correct the WIT type so negative values are rejected at the
boundary before they reach the cast.

fixes dispatch.rs latent cast; aligns inline WIT with hand-rolled records

+++ crates/slicer-host/src/dispatch.rs
@@ -1981,7 +1981,7 @@ fn harvest_paint_segmentation_ir(ctx: wit_host::HostExecutionContext) -> slicer_
 
     let mut per_layer: HashMap<u32, LayerPaintMap> = HashMap::new();
     for (idx, entry) in ctx.paint_region_entries.into_iter().enumerate() {
-        let layer_index = entry.layer_index;
+        let layer_index = entry.layer_index as u32; // validator at wit_host.rs push_paint_region rejects negative
         let layer = per_layer
             .entry(layer_index)
             .or_insert_with(|| LayerPaintMap {
